Remove Gemfile.lock
When developing a gem, use the gemspec method in your Gemfile to avoid
duplication. In general, a gem’s Gemfile should contain the Rubygems source
and a single gemspec line. Do not check your Gemfile.lock into version
control, since it enforces precision that does not exist in the gem
command.

Source: http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/